Refactor: Move message part components to part directory
