+++ components/part/document-preview.tsx
@@ -14,14 +14,14 @@ import { useArtifact } from "@/hooks/use-artifact";
 import type { ArtifactKind } from "@/lib/artifacts/artifact-kind";
 import type { Document } from "@/lib/db/schema";
 import { cn } from "@/lib/utils";
-import type { UIArtifact } from "./artifact";
-import { CodeEditor } from "./code-editor";
+import type { UIArtifact } from "../artifact";
+import { CodeEditor } from "../code-editor";
 import { DocumentToolCall, DocumentToolResult } from "./document";
-import { InlineDocumentSkeleton } from "./document-skeleton";
-import { FileIcon, FullscreenIcon, LoaderIcon, PencilEditIcon } from "./icons";
-import { ImageEditor } from "./image-editor";
-import { SpreadsheetEditor } from "./sheet-editor";
-import { Editor } from "./text-editor";
+import { InlineDocumentSkeleton } from "../document-skeleton";
+import { FileIcon, FullscreenIcon, LoaderIcon, PencilEditIcon } from "../icons";
+import { ImageEditor } from "../image-editor";
+import { SpreadsheetEditor } from "../sheet-editor";
+import { Editor } from "../text-editor";
 
 type DocumentPreviewProps = {
   isReadonly: boolean;
