+++ components/part/generated-image.tsx
